# api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "自定义描述",
        "title": "自定义title",
        "termsOfService": "https://Lirous.com",
        "contact": {
            "name": "Lirous",
            "url": "https://Lirous.com",
            "email": "lirous@lirous.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/Lirous587/go-scaffold/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/captcha": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "生成验证码",
                "parameters": [
                    {
                        "enum": [
                            "image:click"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "WayImageClick"
                        ],
                        "name": "way",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.CaptchaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.invalidParamsResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/v1/captcha/with-answer": {
            "get": {
                "description": "创建新的验证码并返回答案（仅用于测试或开发环境）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "生成带答案的验证码",
                "parameters": [
                    {
                        "enum": [
                            "image:click"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "WayImageClick"
                        ],
                        "name": "way",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.CaptchaAnswerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.invalidParamsResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.VerifyWay": {
            "type": "string",
            "enum": [
                "image:click"
            ],
            "x-enum-varnames": [
                "WayImageClick"
            ]
        },
...
        "handler.CaptchaAnswerResponse": {
            "type": "object",
            "properties": {
                "audio": {
                    "description": "音频验证码",
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "image": {
                    "description": "主图片",
                    "type": "string"
                },
                "thumb": {
                    "description": "缩略图",
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "way": {
                    "$ref": "#/definitions/domain.VerifyWay"
                }
            }
        },
        "handler.CaptchaResponse": {
            "type": "object",
            "properties": {
                "audio": {
                    "description": "音频验证码",
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "image": {
                    "description": "主图片",
                    "type": "string"
                },
                "thumb": {
                    "description": "缩略图",
                    "type": "string"
                }
            }
        },
...
        "response.errorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "response.invalidParamsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string",
                    "example": "invalid params"
                }
            }
        },
        "response.successResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 2000
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "请求成功"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}